+++ src/Network.ts
@@ -18,6 +18,8 @@
  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
+import { SignSchema } from 'nem2-sdk';
+
 // internal dependencies
 import {
     CurveAlgorithm,
@@ -50,7 +52,11 @@ export class Network {
      * @see https://github.com/bitcoinjs/bip32/blob/master/src/bip32.js#L19
      * @var {Network}
      */
-    public static readonly BITCOIN = new Network(0x0488b21e, 0x0488ade4, CurveAlgorithm.secp256k1);
+    public static readonly BITCOIN: Network = new Network(
+        0x0488b21e, // base58 'xpub'
+        0x0488ade4, // base58 'xprv'
+        CurveAlgorithm.secp256k1
+    );
 
     /**
      * CATAPULT protocol extended key prefixes
@@ -59,7 +65,24 @@ export class Network {
      *
      * @var {Network}
      */
-    public static readonly CATAPULT = new Network(0x0488b21e, 0x0488ade4, CurveAlgorithm.ed25519);
+    public static readonly CATAPULT: Network = new Network(
+        0x0488b21e, // base58 'xpub'
+        0x0488ade4, // base58 'xprv'
+        CurveAlgorithm.ed25519
+    );
+
+    /**
+     * CATAPULT public network protocol extended key prefixes
+     *
+     * Result in Base58 notation to `xpub` and `xprv`.
+     *
+     * @var {Network}
+     */
+    public static readonly CATAPULT_PUBLIC: Network = new Network(
+        0x0488b21e, // base58 'xpub'
+        0x0488ade4, // base58 'xprv'
+        CurveAlgorithm.ed25519_keccak
+    );
 
     /**
      * Construct an `Network` object out of its' base58 payload.
@@ -84,5 +107,27 @@ export class Network {
                  */
                 public readonly curve: CurveAlgorithm = CurveAlgorithm.secp256k1) {
 
+    }
+
+    /**
+     * Resolve the signature schema that must be used. This method is only
+     * relevant for Catapult key derivation and will/must not affect BITCOIN
+     * key derivation.
+     *
+     * In case of a `curve` field set to `CurveAlgorithm.ed25519_keccak`, the
+     * underlying SHA3 hasher should use Keccak ; This is to be compatible with
+     * Catapult public network implementations.
+     *
+     * @param   network    {Network}    The network object to interpret
+     * @return  {SignSchema}    Returns the signature schema needed.
+     */
+    public static resolveSignSchema(
+        network: Network
+    ): SignSchema {
+        if (network.curve === CurveAlgorithm.ed25519_keccak) {
+            return SignSchema.KECCAK;
+        }
+
+        return SignSchema.SHA3;
     }
 }